chore: add rudimentary wasi support
This PR adds rudimentary WebAssembly/WASI support
by excluding unsupported calls to non-implemented
WASI/posix syscalls such as setjmp. All
modifications are behind preprocessor directives
to ensure that no other targets are affected by
the patches.

In the future, use of CMake to swap out the needed
source files is a preferred option. For now, this
patch provides the needed functionality while
being as non-destructive as possible.

chore: add build scripts and github action

Adds dockerfile for building in a debian image and
adds a build script for ease of use compiling for
the first time.

A root container is used to install wasi-sdk into
the clang-runtimes directory.

A github workflow is added to smoke test all
pushes and pull requests. Resulting static
library is uploaded to the run as an artifact.