chore: add build scripts and github action
Adds dockerfile for building in a debian image and
adds a build script for ease of use compiling for
the first time.

A root container is used to install wasi-sdk into
the clang-runtimes directory.

A github workflow is added to smoke test all
pushes and pull requests. Resulting static
library is uploaded to the run as an artifact.